added filter code
